fine grained code for TRUNK-3267

+++ api/src/main/java/org/openmrs/api/impl/ObsServiceImpl.java
@@ -12,6 +12,7 @@
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
 package org.openmrs.api.impl;
+
 import java.util.Date;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
@@ -205,36 +206,66 @@ public class ObsServiceImpl extends BaseOpenmrsService implements ObsService {
 	/**
 	 * @see org.openmrs.api.ObsService#purgeObs(org.openmrs.Obs, boolean)
 	 */
-	public void purgeObs(Obs obs, boolean cascade) throws APIException {
-		if (purgeComplexData(obs) == false) {
-			throw new APIException("Unable to purge complex data for obs: " + obs);
-		}
-		
-		if (cascade) {
-			if (obs.isObsGrouping()) {
-				Set<Obs> checkObsGroupMembers = obs.getGroupMembers(true);
-				
-				if (!checkObsGroupMembers.isEmpty()) {
-					for (Obs gMember : checkObsGroupMembers) {
-						
-						Set<Obs> obset = gMember.getRelatedObservations();
-						for (Obs obsMember : obset) {
-							dao.deleteObs(obsMember);
-						}
-					}
-					
-				} else {
-					dao.deleteObs(obs);
-				}
-			} else {
-				dao.deleteObs(obs);
-			}
-			
-		}
-		
-		dao.deleteObs(obs);
-	}
 	
+
+    public void purgeObs(Obs obs, boolean cascade) throws APIException {
+                    if (purgeComplexData(obs) == false) {
+                            throw new APIException("Unable to purge complex data for obs: " + obs);
+                    }
+                   
+                    if (cascade) {
+                //check for related Obs
+               
+                            //check for parent Obs
+                            if (obs.isObsGrouping()) {
+                                    //check for child Obs
+                                    Set<Obs> checkObsGroupMembers = obs.getGroupMembers(true);
+                                   
+                                    if (!checkObsGroupMembers.isEmpty()) {
+                                            for (Obs gMember : checkObsGroupMembers) {
+                                                   
+                                                    //check to see if Obs is a parent
+                                                    if (gMember.isObsGrouping()) {
+                                                            Set<Obs> obset = gMember.getGroupMembers();
+                                                            for (Obs obsMember : obset) {
+                                                                    dao.deleteObs(obsMember);
+                                                            }
+                                                           
+                                                            if (gMember != null)
+                                                                    dao.deleteObs(gMember);
+                                                           
+                                                    } else { //if it has no child
+                               
+                                for(Obs gMemberRelObs:gMember.getRelatedObservations()){
+                                    if(gMemberRelObs !=null){
+                                        dao.deleteObs(gMemberRelObs);
+                                    }
+     
+                                }
+                               
+                                                    }
+                                                   
+                                            }
+                                           
+                                    } else {
+                                            if (obs != null)
+                                                    dao.deleteObs(obs);
+                                    }
+                            } else { //if Obs is not obsGroup
+                   
+                   for(Obs relObs:obs.getRelatedObservations()){
+                       if(relObs !=null){
+                           dao.deleteObs(obs);  
+                       }
+                     
+                   }        
+                            }
+                           
+                    }     
+                    dao.deleteObs(obs);
+            }
+
+
 	/**
 	 * @see org.openmrs.api.ObsService#purgeObs(org.openmrs.Obs)
 	 */
